requirements.txt
numpy==1.26.2
Pillow==10.1.0
pydicom==2.4.3

shapely~=2.0.2